Extract field definitions
