+++ controllers/apply/contacts-next/fields/main-contact-phone.js
@@ -1,0 +1,9 @@
+'use strict';
+const PhoneField = require('../../lib/field-types/phone');
+
+module.exports = function(locale) {
+    return new PhoneField({
+        locale: locale,
+        name: 'mainContactPhone'
+    });
+};
